@@ -3,10 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-scripts": "^5.0.1",
-    "react-transition-group": "^4.4.5"
+    "react-transition-group": "^2.2.1"
   },
   "scripts": {
     "start": "react-scripts start",
