@@ -1,5 +1,6 @@
 import React, { Component } from "react";
 import Transition from "react-transition-group/Transition";
+
 import "./App.css";
 import Modal from "./components/Modal/Modal";
 import Backdrop from "./components/Backdrop/Backdrop";
@@ -25,18 +26,24 @@ class App extends Component {
         <h1>React Animations</h1>
         <button
           className="Button"
-          onClick={() => {
-            this.setState((prevState) => ({ showBlock: !prevState.showBlock }));
-          }}
+          onClick={() =>
+            this.setState((prevState) => ({ showBlock: !prevState.showBlock }))
+          }
         >
           Toggle
         </button>
         <br />
         <Transition
+          in={this.state.showBlock}
+          timeout={1000}
           mountOnEnter
           unmountOnExit
-          in={this.state.showBlock}
-          timeout={300}
+          onEnter={() => console.log("onEnter")}
+          onEntering={() => console.log("onEntering")}
+          onEntered={() => console.log("onEntered")}
+          onExit={() => console.log("onExit")}
+          onExiting={() => console.log("onExiting")}
+          onExited={() => console.log("onExited")}
         >
           {(state) => (
             <div
@@ -48,18 +55,12 @@ class App extends Component {
                 transition: "opacity 1s ease-out",
                 opacity: state === "exiting" ? 0 : 1,
               }}
-            >
-              Block
-            </div>
+            />
           )}
         </Transition>
-
         <Modal show={this.state.modalIsOpen} closed={this.closeModal} />
-
-        {this.state.modalIsOpen ? (
-          <Backdrop show={this.state.modalIsOpen} />
-        ) : null}
-        <button onClick={this.showModal} className="Button">
+        {this.state.modalIsOpen ? <Backdrop show /> : null}
+        <button className="Button" onClick={this.showModal}>
           Open Modal
         </button>
         <h3>Animating Lists</h3>
