@@ -7,10 +7,8 @@ const backdrop = (props) => {
     "Backdrop",
     props.show ? "BackdropOpen" : "BackdropClosed",
   ];
-  return (
-    <>
-      <div className={cssClasses.join(" ")}></div>
-    </>
-  );
+
+  return <div className={cssClasses.join(" ")}></div>;
 };
+
 export default backdrop;
